+++ R/intervals.R
@@ -26,15 +26,34 @@ head.intervals = function(x, n) make_intervals(head(x$start, n), head(x$end, n))
 
 tail.intervals = function(x, n) make_intervals(tail(x$start, n), tail(x$end, n))
 
+c.intervals = function(...) {
+	dots = list(...)
+	start = do.call(c, lapply(dots, function(x) x$start))
+	end = do.call(c, lapply(dots, function(x) x$end))
+	make_intervals(start, end)
+}
+
 `[.intervals` = function(x, i, ...) {
 	make_intervals(x$start[i], x$end[i])
 }
 
 format.intervals = function(x, ...) {
+	mformat = function(x, ..., digits = getOption("digits")) {
+		if (inherits(x, "PCICt")) 
+			format(x, ...)
+		else
+			format(x, digits = digits, ...) 
+	}
 	if (inherits(x$start, "units")) {
 		stopifnot(units(x$start) == units(x$end))
 		paste0("[", format(as.numeric(x$start), ...), ",", format(as.numeric(x$end), ...), ") ",
 			"[", as.character(units(x$start)), "]")
 	} else
-		paste0("[", format(x$start, ...), ",", format(x$end, ...), ")")
+		paste0("[", mformat(x$start, ...), ",", mformat(x$end, ...), ")")
+}
+
+find_interval = function(x, intervals) {
+	i = findInterval(x, c(intervals$start, tail(intervals$end, 1)))
+	i[i == 0 | i > length(intervals)] = NA
+	i
 }