+++ R/ops.R
@@ -75,6 +75,8 @@ st_apply = function(X, MARGIN, FUN, ...) UseMethod("st_apply")
 #' @param ... arguments passed on to \code{FUN}
 #' @param CLUSTER cluster to use for parallel apply; see \link[parallel]{makeCluster}
 #' @param PROGRESS logical; if \code{TRUE}, use \code{pbapply::pbapply} to show progress bar
+#' @param FUTURE logical;if \code{TRUE}, use \code{future.apply::future_apply} 
+#' @param rename logical; if \code{TRUE} and \code{X} has only one attribute and \code{FUN} is a simple function name, rename the attribute of the returned object to the function name
 #' @return object of class \code{stars} with accordingly reduced number of dimensions; in case \code{FUN} returns more than one value, a new dimension is created carrying the name of the function used; see the examples.
 #' @examples
 #' tif = system.file("tif/L7_ETMs.tif", package = "stars")
@@ -84,8 +86,12 @@ st_apply = function(X, MARGIN, FUN, ...) UseMethod("st_apply")
 #' st_apply(x, 3, mean)   # mean of all pixels for each band
 #' st_apply(x, "band", mean) # equivalent to the above
 #' st_apply(x, 1:2, range) # min and max band value for each pixel
+#' # to get a progress bar also in non-interactive mode, specify:
+#' if (require(pbapply)) { # install it, if FALSE
+#'   pboptions(type = "timer")
+#' }
 #' @export
-st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE) {
+st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE, FUTURE = FALSE, rename = TRUE) {
 	fname <- paste(deparse(substitute(FUN), 50), collapse = "\n")
 	if (is.character(MARGIN))
 		MARGIN = match(MARGIN, names(dim(X)))
@@ -93,15 +99,22 @@ st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE)
 
 	if (PROGRESS && !requireNamespace("pbapply", quietly = TRUE))
 		stop("package pbapply required, please install it first")
+	
+	if (FUTURE && !requireNamespace("future.apply", quietly = TRUE))
+	  stop("package future.apply required, please install it first")
 
 	fn = function(y, ...) {
 		ret = if (PROGRESS)
-				pbapply::pbapply(y, MARGIN, FUN, ..., cl = CLUSTER)
+				pbapply::pbapply(X = y, MARGIN = MARGIN, FUN = FUN, ..., cl = CLUSTER)
 			else {
-				if (is.null(CLUSTER))
-					apply(y, MARGIN, FUN, ...)
-				else
-					parallel::parApply(CLUSTER, y, MARGIN, FUN, ...)
+				if (is.null(CLUSTER) && !FUTURE)
+					apply(X = y, MARGIN = MARGIN, FUN = FUN, ...)
+				else if (FUTURE) {
+					oopts = options(future.globals.maxSize = +Inf)
+					on.exit(options(oopts))
+					future.apply::future_apply(y, MARGIN = MARGIN, FUN = FUN, ...)
+				} else
+					parallel::parApply(CLUSTER, X = y, MARGIN = MARGIN, FUN = FUN, ...)
 			}
 		if (is.array(ret))
 			ret
@@ -110,9 +123,11 @@ st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE)
 	}
 	ret = lapply(X, fn, ...) 
 	dim_ret = dim(ret[[1]])
-	if (length(dim_ret) == length(MARGIN)) # FUN returned a single value
+	ret = if (length(dim_ret) == length(MARGIN)) { # FUN returned a single value
+		if (length(ret) == 1 && rename && make.names(fname) == fname)
+			ret = setNames(ret, fname)
 		st_stars(ret, st_dimensions(X)[MARGIN])
-	else { # FUN returned multiple values:
+	} else { # FUN returned multiple values: need to set dimension name & values
 		orig = st_dimensions(X)[MARGIN]
 		r = attr(orig, "raster")
 		dims = c(structure(list(list()), names = fname), orig)
@@ -122,4 +137,7 @@ st_apply.stars = function(X, MARGIN, FUN, ..., CLUSTER = NULL, PROGRESS = FALSE)
 				create_dimension(to = dim_ret[1])
 		st_stars(ret, dimensions = create_dimensions(dims, r))
 	}
+	for (i in seq_along(ret))
+		names(dim(ret[[i]])) = names(st_dimensions(ret))
+	ret
 }