+++ R/proxy.R
@@ -110,17 +110,26 @@ fetch = function(x, downsample = 0, ...) {
 	}
 	rasterio = list(nXOff = dx$from, nYOff = dy$from, nXSize = nXSize, nYSize = nYSize, 
 		nBufXSize = nBufXSize, nBufYSize = nBufYSize)
-	if (!is.null(bands <- d[["band"]])) # we may want to select here
-		rasterio$bands = bands$values %||% bands$from:bands$to
+	if (!is.null(bands <- d[["band"]]) && !is.null(bands$values) && is.numeric(bands$values)) # we want to select here
+		rasterio$bands = bands$values
 
 	# do it:
 	ret = lapply(x, read_stars, RasterIO = rasterio, 
 		NA_value = attr(x, "NA_value") %||% NA_real_, ...)
 
-	if (length(ret) == 1)
+	ret = if (length(ret) == 1)
 		st_redimension(ret[[1]])
 	else
 		do.call(c, lapply(ret, st_redimension))
+	
+	new_dim = st_dimensions(ret)
+	for (dm in setdiff(names(d), xy)) # copy over non x/y dimension values, if present:
+		if (dm %in% names(new_dim))
+			new_dim[[dm]] = d[[dm]]
+#		if (!is.null(v <- d[[dm]]$values))
+#			new_dim[[dm]]$values = v
+
+	st_stars(setNames(ret, names(x)), new_dim)
 }
 
 check_xy_warn = function(call, dimensions) {
@@ -214,27 +223,35 @@ aperm.stars_proxy = function(a, perm = NULL, ...) {
 	collect(a, match.call(), "aperm", "a")
 }
 
+
 #' @export
-"[.stars_proxy" = function(x, ..., drop = FALSE, crop = TRUE) {
+merge.stars_proxy = function(x, y, ...) {
+	collect(x, match.call(), "merge")
+}
+
+
+#' @export
+"[.stars_proxy" = function(x, i = TRUE, ..., drop = FALSE, crop = TRUE) {
 	mc = match.call()
 	lst = as.list(mc)
 	if (length(lst) < 3)
 		return(x) # 
-	if (as.character(lst[[3]]) != "" && crop) { # i present: do attr selection or bbox now:
-		x = if (is.character(lst[[3]]))
-			st_stars_proxy(unclass(x)[ lst[[3]] ], st_dimensions(x))
-		else {
-			i = as.character(lst[[3]])
-			if (inherits(get(i), c("sf", "sfc", "stars", "bbox")))
-				st_crop(x, get(i), ...)
-			else
-				stop(paste("unrecognized selector:", i))
-		}
-		if (length(lst) == 3 && crop) # we're done
-			return(x)
-		lst[[3]] = TRUE # this one has been handled
+	if (missing(i)) # insert:
+		lst = c(lst[1:2], i = TRUE, lst[-(1:2)])
+	if (inherits(i, c("character", "logical", "numeric"))) {
+		x = st_stars_proxy(unclass(x)[ lst[[3]] ], st_dimensions(x))
+		lst[["i"]] = TRUE # this one has been handled now
+	} else if (crop && inherits(i, c("sf", "sfc", "stars", "bbox"))) {
+		x = st_crop(x, i, ...) # does bounding box cropping only
+		if (inherits(i, c("stars", "bbox")))
+			lst[["i"]] = TRUE # this one has been handled now
 	}
-	collect(x, as.call(lst), "[") # postpone every arguments > 3 to after reading cells
+
+	# return:
+	if (length(lst) == 3 && isTRUE(lst[["i"]])) 
+		x
+	else # still processing the geometries inside the bbox:
+		collect(x, as.call(lst), "[") # postpone every arguments > 3 to after reading cells
 }
 
 # shrink bbox with e * width in each direction
@@ -265,7 +282,7 @@ st_crop.stars_proxy = function(x, y, ..., crop = TRUE, epsilon = 0) {
 		if (epsilon != 0)
 			bb = bb_shrink(bb, epsilon)
 		# FIXME: document how EXACTLY cropping works; https://github.com/hypertidy/tidync/issues/73
-		cr = round(colrow_from_xy(matrix(bb, 2, byrow=TRUE), get_geotransform(dm)) + 0.5)
+		cr = colrow_from_xy(matrix(bb, 2, byrow=TRUE), dm)
 		for (i in seq_along(dm)) {
 			if (names(d[i]) == xd) {
 				dm[[ xd ]]$from = max(1, cr[1, 1])
@@ -285,6 +302,16 @@ st_crop.stars_proxy = function(x, y, ..., crop = TRUE, epsilon = 0) {
 #' @export
 st_apply.stars_proxy = function(X, MARGIN, FUN, ...) {
 	collect(X, match.call(), "st_apply", "X")
+}
+
+#' @export
+predict.stars_proxy = function(object, model, ...) {
+	collect(object, match.call(), "predict", "object")
+}
+
+#' @export
+split.stars_proxy = function(x, ...) {
+	collect(x, match.call(), "split")
 }
 
 #nocov start