+++ tests/stars.R
@@ -1,3 +1,4 @@
+Sys.setenv(TZ="UTC")
 suppressPackageStartupMessages(library(stars))
 set.seed(13521) # runif
 tif = system.file("tif/L7_ETMs.tif", package = "stars")
@@ -38,6 +39,7 @@ nc = system.file("nc/tos_O1_2001-2002.nc", package = "stars")
 st_as_stars(st_bbox(x))
 st_as_stars(st_bbox(x), deltax = 20, deltay = 20)
 df = as.data.frame(x)
+units::drop_units(x)
 
 dimnames(x)
 dimnames(x) <- letters[1:3]