+++ NEWS.md
@@ -1,3 +1,9 @@
+# version 0.3-1
+
+* handle bounds, when given, in `read_ncdf`
+
+* provide time parsing (POSIXct, PCICt) for `read_ncdf`; #115
+
 # version 0.3-0
 
 * add `st_area` method to return raster grid cell sizes; #99